Move generateUserTokens to utils file
